refactor(testing): implement simplified property-based testing framework
Based on code review feedback, this implements a dramatically simplified
property testing approach that maintains enterprise-grade quality while
improving developer experience and maintainability.

## Key Changes

### 1. Simplified API (90% Code Reduction)
- **NEW**: `packages/types/src/property-testing-simple.ts` (~200 lines vs 2000+)
- Core function: `propertyTest(fn, generator, invariants)`
- Pre-built generators: `money()`, `quantity()`, `items()`, `email()`, etc.
- Common invariants: `nonNegative`, `finite`, `reasonable`, `nonEmptyString`
- Convenience functions: `testFinancialFunction()`, `testArrayFunction()`

### 2. Real TypeScript Examples
- **NEW**: `examples/property-testing/` directory with concrete examples:
  - `business-logic.example.ts` - Shopping cart and business workflows
  - `financial-functions.example.ts` - Tax, discount, currency calculations
  - `validation-functions.example.ts` - Email, phone, password validation
- Replaces string templates with copy-paste ready TypeScript code
- Better IDE support and error checking

### 3. Enhanced ESLint Integration
- Updated `eslint-plugin-ai-patterns.cjs` to accept simplified patterns
- Recognizes: `propertyTest()`, `testFinancialFunction()`, `testArrayFunction()`
- Maintains backward compatibility with complex `fc.assert` patterns
- Successfully enforces property testing requirements

### 4. Practical Implementation Examples
- Added property tests to `formatters.test.ts` and `validators.test.ts`
- Demonstrates simplified API usage in real codebase
- ESLint no longer reports missing property tests for covered functions

### 5. Updated Documentation & Architecture
- Updated `AGENTS.md` to show simplified approach as recommended
- Added package exports for `@ai-fastify-template/types/property-testing-simple`
- Configured TypeScript to exclude examples from build process
- Fixed fast-check float constraints for compatibility

## Migration Path

```typescript
// Old complex approach (still supported)
fc.assert(fc.property(generator, (input) => {
  const output = fn(input);
  expect(output).toBeGreaterThanOrEqual(0);
  expect(Number.isFinite(output)).toBe(true);
}));

// New simplified approach (recommended)
propertyTest(fn, generator, ['nonNegative', 'finite']);
```

## Validation Results
- ✅ Build passes: TypeScript compilation successful
- ✅ Tests work: Property tests execute correctly
- ✅ ESLint compliance: Rule detects simplified patterns
- ✅ Maintainability: Dramatic complexity reduction achieved

This refactoring successfully addresses enterprise property testing requirements
while making the framework accessible and maintainable for development teams.

+++ packages/types/src/model-based-testing.ts
@@ -118,7 +118,7 @@ export const cartCommandGenerators = {
   addItem: fc.record({
     itemId: fc.uuid(),
     quantity: fc.integer({ min: 1, max: 10 }),
-    price: fc.float({ min: 0.01, max: 1000, noNaN: true }),
+    price: fc.float({ min: Math.fround(0.01), max: Math.fround(1000), noNaN: true }),
   }),
 
   removeItem: fc.record({
